round 3 simplifications: internalize cycle_nonce, drop newtype, merge runner
Three convergent simplifications:

- cycle_nonce moved out of CycleConfig and into run_cycle. Both callers
  (sidecar daemon, restart helper) were computing it identically as
  base.checkpoint_id as u128 and pre-fetching latest_checkpoint just to
  derive it. run_cycle already resolves the base in Phase 1; the
  derivation is now there. Drops the field from CycleConfig, the pre-
  fetch in both callers, and the speculative "plus attempt counter"
  comment that no caller implemented.

- PeerResponses<T> newtype replaced with Vec<T>. The struct had a single
  public field and every consumer immediately reached through .responses.
  Round 2's "self-documenting at call sites" argument didn't hold up
  given .responses.iter() reads everywhere.

- restart.rs + sidecar.rs merged into runner.rs. Both files were small
  (~150 lines each) doing the same wire-stages-and-call-run_cycle work
  with two different (materializer, finalizer, min_mutations) tuples.
  One file, two functions, less module fragmentation.

26 unit tests + 4 DB-dependent tests passing. fmt + clippy clean.

After three simplification rounds the protocol module is ~470 lines
lighter than the pre-loop state. The 5-trait split (Materializer,
TerminalAction, ConsensusTransport, MutationStore, GraphHasher) survived
intact — each carries real test-seam value.
